PM-4308 #time 2h ensure user can't manually close ai screening phase when there are pending decissions
