+++ src/common/challenge-helper.js
@@ -449,7 +449,8 @@ class ChallengeHelper {
       config.REVIEWS_API_URL || "https://api.topcoder-dev.com",
       "/"
     );
-    const url = `${reviewsApiBaseUrl}/v6/ai-review/configs`;
+    
+    const url = `${reviewsApiBaseUrl}/ai-review/configs`;
 
     for (const aiReviewConfig of aiReviewConfigs) {
       const payload = {
@@ -491,10 +492,10 @@ class ChallengeHelper {
   async getAIReviewTemplateById(templateId) {
     const token = await getM2MToken();
     const reviewsApiBaseUrl = _.trimEnd(
-      config.REVIEWS_API_URL || "https://api.topcoder-dev.com",
+      config.REVIEWS_API_URL || "http://localhost:3002",
       "/"
     );
-    const url = `${reviewsApiBaseUrl}/v6/ai-review/templates/${templateId}`;
+    const url = `${reviewsApiBaseUrl}/ai-review/templates/${templateId}`;
 
     try {
       const response = await axios.get(url, {
